add pyrefly type checking and pipeline migration specs (#103)
* add pyrefly type checker and fix all type errors

- Add pyrefly to requirements.txt and configure in pyproject.toml with
  search_path pointing at scripts/ so lib.* imports resolve
- Fix missing `from pathlib import Path` in fix-media-timestamp.py (real
  bug — would crash if EXIF read error path was hit)
- Add else-branch stubs for conditional darwin imports in
  fix-media-timestamp.py so pyrefly can resolve the names on all platforms
- Add assert after _ensure_running() in exiftool.py to prove _process is
  not None
- Guard `"task" in tasks` checks in media-pipeline.py with `tasks and`
  for None safety (tasks param typed as `set | None`)
- Fix `format_bytes` param type from int to float (true division produces
  float)
- Add explicit `Optional[str]` annotation on archive_dir in import-media.py
  to break pyrefly dependency cycle inference
- Initialize current_birth_dt before conditional branch in
  fix-media-timestamp.py to satisfy unbound-name check
- Suppress 2 macOS-specific attributes (st_birthtime, st_flags) that are
  guarded by try/except AttributeError at runtime

All 210 tests pass; pyrefly check reports 0 errors.


* add pipeline migration specs and fix import side-effects

Move signal.signal(SIGINT, ...) from module level to main() in
tag-media.py and fix-media-timestamp.py. Module-level signal
registration is a side-effect that overrides the caller's handler on
import — blocking the subprocess-to-module-call migration.

Add three ordered specs for the pipeline internal migration:
- typed-return-contracts: replace @@key=value prints inside functions
  with dataclass returns. @@ moves to pipeline boundary only.
- subprocess-to-module-calls: import and call script functions directly
  instead of spawning Python subprocesses (5 remain, 1 already done).
- jsonl-app-ipc: replace @@key=value at the process boundary with
  structured JSONL. Schema-defined events, backward-compatible
  transition via --output-format flag.

Each spec enables the next. Together they lay the foundation for the
time-correction-pipeline-step spec (already exists, unchanged).

Update TODO.md with the four-step migration order.


* fix @@ lifecycle in migration specs

Clarify the @@ deletion path — each step must maintain app
functionality:

- Step 2 (typed returns): @@ prints deleted from function bodies;
  main() temporarily serialises the returned dataclass to @@ so
  subprocess callers (media-pipeline) still work
- Step 3 (module calls): pipeline calls functions directly; @@
  serialisation in scripts' main() is dead code and deleted.
  Pipeline's emit() still prints @@ for the macOS app.
- Step 4 (JSONL): atomic switchover — pipeline emits JSON, app
  parser switches to JSONDecoder, @@ in emit() deleted.
  Dual-format flag during transition if needed.


---------

+++ scripts/backup-source-video-to-nas.py
@@ -68,7 +68,7 @@ def transform_local_to_remote_path(local_path: str, local_base_path: str, remote
     relative_path = local_path[len(local_base_path):]
     return os.path.join(remote_base_path, relative_path)
 
-def format_bytes(num_bytes: int) -> str:
+def format_bytes(num_bytes: float) -> str:
     """Format bytes as human-readable string"""
     for unit in ['B', 'KB', 'MB', 'GB', 'TB']:
         if abs(num_bytes) < 1024:
